#488: refactored everything but tests
