+++ src/backend/src/utils/projects.utils.ts
@@ -1,21 +1,8 @@
-import { Prisma, WBS_Element_Status } from '@prisma/client';
+import { WBS_Element_Status } from '@prisma/client';
 import prisma from '../prisma/prisma';
-import {
-  Project,
-  WbsElementStatus,
-  calculateEndDate,
-  calculateProjectEndDate,
-  calculatePercentExpectedProgress,
-  calculateTimelineStatus,
-  calculateDuration,
-  calculateProjectStartDate
-} from 'shared';
-import { descBulletConverter, wbsNumOf } from './utils';
-import riskQueryArgs from '../prisma-query-args/risks.query-args';
-import riskTransformer from '../transformers/risks.transformer';
+import { WbsElementStatus } from 'shared';
 import { buildChangeDetail } from '../utils/utils';
-import { calculateWorkPackageProgress } from './work-packages.utils';
-import userTransformer from '../transformers/user.transformer';
+import { NotFoundException } from './errors.utils';
 
 /**
  * calculate the project's status based on its workpacakges' status
@@ -26,170 +13,8 @@ export const calculateProjectStatus = (proj: { workPackages: { wbsElement: { sta
   if (proj.workPackages.length === 0) return WbsElementStatus.Inactive;
 
   if (proj.workPackages.every((wp) => wp.wbsElement.status === WbsElementStatus.Complete)) return WbsElementStatus.Complete;
-  else if (proj.workPackages.findIndex((wp) => wp.wbsElement.status === WbsElementStatus.Active) !== -1)
-    return WbsElementStatus.Active;
+  else if (proj.workPackages.some((wp) => wp.wbsElement.status === WbsElementStatus.Active)) return WbsElementStatus.Active;
   return WbsElementStatus.Inactive;
-};
-
-export const manyRelationArgs = Prisma.validator<Prisma.ProjectArgs>()({
-  include: {
-    wbsElement: {
-      include: {
-        projectLead: true,
-        projectManager: true,
-        changes: { include: { implementer: true } }
-      }
-    },
-    team: true,
-    goals: { where: { dateDeleted: null } },
-    features: { where: { dateDeleted: null } },
-    otherConstraints: { where: { dateDeleted: null } },
-    risks: { where: { dateDeleted: null }, ...riskQueryArgs },
-    workPackages: {
-      where: {
-        wbsElement: {
-          dateDeleted: null
-        }
-      },
-      include: {
-        wbsElement: {
-          include: {
-            projectLead: true,
-            projectManager: true,
-            changes: { include: { implementer: true } }
-          }
-        },
-        dependencies: true,
-        expectedActivities: true,
-        deliverables: true
-      }
-    }
-  }
-});
-
-export const uniqueRelationArgs = Prisma.validator<Prisma.WBS_ElementArgs>()({
-  include: {
-    project: {
-      include: {
-        team: true,
-        goals: { where: { dateDeleted: null } },
-        features: { where: { dateDeleted: null } },
-        otherConstraints: { where: { dateDeleted: null } },
-        risks: { where: { dateDeleted: null }, ...riskQueryArgs },
-        workPackages: {
-          where: {
-            wbsElement: {
-              dateDeleted: null
-            }
-          },
-          include: {
-            wbsElement: {
-              include: {
-                projectLead: true,
-                projectManager: true,
-                changes: { include: { implementer: true } }
-              }
-            },
-            dependencies: true,
-            expectedActivities: true,
-            deliverables: true
-          }
-        }
-      }
-    },
-    projectLead: true,
-    projectManager: true,
-    changes: { include: { implementer: true } }
-  }
-});
-
-export const projectTransformer = (
-  payload: Prisma.ProjectGetPayload<typeof manyRelationArgs> | Prisma.WBS_ElementGetPayload<typeof uniqueRelationArgs>
-): Project => {
-  const wbsElement = 'wbsElement' in payload ? payload.wbsElement : payload;
-  const project = 'project' in payload ? payload.project! : payload;
-  const wbsNum = wbsNumOf(wbsElement);
-  let team = undefined;
-  if (project.team) {
-    team = {
-      teamId: project.team.teamId,
-      teamName: project.team.teamName
-    };
-  }
-  const { projectLead, projectManager } = wbsElement;
-
-  return {
-    id: project.projectId,
-    wbsNum,
-    dateCreated: wbsElement.dateCreated,
-    name: wbsElement.name,
-    status: calculateProjectStatus(project),
-    projectLead: projectLead ? userTransformer(projectLead) : undefined,
-    projectManager: projectManager ? userTransformer(projectManager) : undefined,
-    changes: wbsElement.changes.map((change) => ({
-      changeId: change.changeId,
-      changeRequestId: change.changeRequestId,
-      wbsNum,
-      implementer: userTransformer(change.implementer),
-      detail: change.detail,
-      dateImplemented: change.dateImplemented
-    })),
-    team,
-    summary: project.summary,
-    budget: project.budget,
-    gDriveLink: project.googleDriveFolderLink ?? undefined,
-    taskListLink: project.taskListLink ?? undefined,
-    slideDeckLink: project.slideDeckLink ?? undefined,
-    bomLink: project.bomLink ?? undefined,
-    rules: project.rules,
-    duration: calculateDuration(project.workPackages),
-    startDate: calculateProjectStartDate(project.workPackages),
-    endDate: calculateProjectEndDate(project.workPackages),
-    goals: project.goals.map(descBulletConverter),
-    features: project.features.map(descBulletConverter),
-    otherConstraints: project.otherConstraints.map(descBulletConverter),
-    risks: project.risks.map(riskTransformer),
-    workPackages: project.workPackages.map((workPackage) => {
-      const endDate = calculateEndDate(workPackage.startDate, workPackage.duration);
-      const progress = calculateWorkPackageProgress(workPackage.deliverables, workPackage.expectedActivities);
-      const expectedProgress = calculatePercentExpectedProgress(
-        workPackage.startDate,
-        workPackage.duration,
-        workPackage.wbsElement.status
-      );
-
-      return {
-        id: workPackage.workPackageId,
-        wbsNum: wbsNumOf(workPackage.wbsElement),
-        dateCreated: workPackage.wbsElement.dateCreated,
-        name: workPackage.wbsElement.name,
-        status: workPackage.wbsElement.status as WbsElementStatus,
-        projectLead: workPackage.wbsElement.projectLead ? userTransformer(workPackage.wbsElement.projectLead) : undefined,
-        projectManager: workPackage.wbsElement.projectManager
-          ? userTransformer(workPackage.wbsElement.projectManager)
-          : undefined,
-        changes: workPackage.wbsElement.changes.map((change) => ({
-          changeId: change.changeId,
-          changeRequestId: change.changeRequestId,
-          wbsNum: wbsNumOf(workPackage.wbsElement),
-          implementer: userTransformer(change.implementer),
-          detail: change.detail,
-          dateImplemented: change.dateImplemented
-        })),
-        orderInProject: workPackage.orderInProject,
-        progress,
-        startDate: workPackage.startDate,
-        endDate,
-        duration: workPackage.duration,
-        expectedProgress,
-        timelineStatus: calculateTimelineStatus(progress, expectedProgress),
-        dependencies: workPackage.dependencies.map(wbsNumOf),
-        expectedActivities: workPackage.expectedActivities.map(descBulletConverter),
-        deliverables: workPackage.deliverables.map(descBulletConverter),
-        projectName: wbsElement.name
-      };
-    })
-  };
 };
 
 // gets highest current project number
@@ -296,6 +121,6 @@ export const createRulesChangesJson = (
 export const getUserFullName = async (userId: number | null): Promise<string | null> => {
   if (!userId) return null;
   const user = await prisma.user.findUnique({ where: { userId } });
-  if (!user) throw new Error('user not found');
+  if (!user) throw new NotFoundException('User', userId);
   return `${user.firstName} ${user.lastName}`;
 };
