+++ src/frontend/src/pages/CreateProjectPage/CreateProjectForm.tsx
@@ -24,24 +24,25 @@ const CreateProjectForm: React.FC = () => {
 
   if (isLoading || !auth.user) return <LoadingIndicator />;
 
-  const { userId } = auth.user;
-
   const handleCancel = () => history.goBack();
 
   const handleSubmit = async (project: CreateProjectFormInputs) => {
     const { name, carNumber, crId, summary } = project;
 
     const payload = {
-      userId,
       crId,
       name,
       carNumber,
       summary
     };
 
-    const createdWbsNum = await mutateAsync(payload);
+    try {
+      const createdWbsNum = await mutateAsync(payload);
 
-    history.push(`${routes.PROJECTS}/${createdWbsNum}`);
+      history.push(`${routes.PROJECTS}/${createdWbsNum}`);
+    } catch (error) {
+      alert(error);
+    }
   };
 
   return <CreateProjectFormView onCancel={handleCancel} onSubmit={handleSubmit} allowSubmit={auth.user.role !== 'GUEST'} />;
